Current MINCImageIO can't be reused for writing.
The current implementation of MINCImageIO keeps state of the image that it has
read and does not work properly if we attempt to reuse the same object
instantiation for writting. As a temporary fix in the test here, we create a
second MINCImageIO instance and provide it to the writer. The real solution is
to fix the state keeping that currently exist in the MINCImageIO class.